--- a/src/core/model_standard.jl
+++ b/src/core/model_standard.jl
@@ -276,3 +276,6 @@ end
 schedulername(x::Union{Function,DataType}) = nameof(x)
 schedulername(x) = Symbol(typeof(x))
 
+function remove_all_from_model!(model::StandardABM)
+    empty!(agent_container(model))
+end