--- a/src/core/space_interaction_API.jl
+++ b/src/core/space_interaction_API.jl
@@ -446,4 +446,3 @@ function resorvoir_sampling_single(iter, model)
         w *= rand(rng)
     end
 end
-