--- a/src/spaces/grid_multi.jl
+++ b/src/spaces/grid_multi.jl
@@ -263,3 +263,4 @@ ids_in_position(pos::ValidPos, model::ABM{<:GridSpace}) = ids_in_position(pos, a
 function ids_in_position(pos::ValidPos, space::GridSpace)
     return space.stored_ids[pos...]
 end
+