--- a/test/api_tests.jl
+++ b/test/api_tests.jl
@@ -135,14 +135,13 @@ end
     @test nagents(model) == 3
 
     model = StandardABM(GridAgent{2}, GridSpace((10, 10)), warn_deprecation = false)
-
     # Testing remove_all!(model::ABM)
     for i in 1:20
         add_agent_single!(GridAgent{2}, model)
     end
     remove_all!(model)
     @test nagents(model) == 0
-    @test all(isempty(p) for p in positions(model))
+    @test all(p -> isempty(p, model), positions(model)) == true
 
     # Testing remove_all!(model::ABM, n::Int)
     for i in 1:20
@@ -162,22 +161,20 @@ end
     @test nagents(model) == 5
 
     model = StandardABM(GridAgent{2}, GridSpaceSingle((10, 10)), warn_deprecation = false)
-
     for i in 1:20
         add_agent_single!(GridAgent{2}, model)
     end
     remove_all!(model)
     @test nagents(model) == 0
-    @test all(isempty(p) for p in positions(model))
+    @test all(p -> isempty(p, model), positions(model)) == true
 
-    model = StandardABM(GridAgent{2}, GridSpaceSingle((10, 10)), warn_deprecation = false)
-
+    model = StandardABM(ContinuousAgent{2, Float64}, ContinuousSpace((10, 10)), warn_deprecation = false)
     for i in 1:20
-        add_agent_single!(GridAgent{2}, model)
+        add_agent!(ContinuousAgent{2, Float64}, model, SVector(10*rand(), 10*rand()))
     end
     remove_all!(model)
     @test nagents(model) == 0
-    @test all(isempty(p) for p in positions(model))
+    @test all(p -> isempty(p, abmspace(model).grid), positions(abmspace(model).grid)) == true
 
 end