--- a/test/randomness_tests.jl
+++ b/test/randomness_tests.jl
@@ -17,66 +17,63 @@ using StableRNGs
 end
 
 @testset "sample!" begin
-    @testset "container=$(c)" for c in (Vector, Dict)
-        rng = StableRNG(50)
-        model4 = StandardABM(Agent1, GridSpace((2, 2)); rng = rng, container = c, warn_deprecation = false)
-        add_agent!((1,1), Agent1, model4)
-        add_agent!((2,2), Agent1, model4)
-        sample!(model4, 4)
-        agents = [Agent1(4, (2, 2)), Agent1(2, (2, 2)), Agent1(3, (2, 2)), Agent1(1, (1, 1))]
-        res = c isa Dict ? Dict(a.id => a for a in agents) : agents
-        res_fields = [getfield(res[k], f) for f in fieldnames(Agent1) for k in keys(res)]
-        agents_fields = [getfield(a, f) for f in fieldnames(Agent1) for a in allagents(model4)]
-        @test sort(collect(allids(model4))) == sort(keys(res))
-        @test sort(res_fields) == sort(agents_fields)
-        sample!(model4, 2)
-        agents = [Agent1(4, (2, 2)), Agent1(1, (1, 1))]
-        res = c isa Dict ? Dict(a.id => a for a in agents) : agents
-        res_fields = [getfield(res[k], f) for f in fieldnames(Agent1) for k in keys(res)]
-        agents_fields = [getfield(a, f) for f in fieldnames(Agent1) for a in allagents(model4)]
-        @test sort(collect(allids(model4))) == sort(keys(res))
-        @test sort(res_fields) == sort(agents_fields)
+    rng = StableRNG(50)
+    model4 = StandardABM(Agent1, GridSpace((2, 2)); rng = rng, warn_deprecation = false)
+    add_agent!((1,1), Agent1, model4)
+    add_agent!((2,2), Agent1, model4)
+    sample!(model4, 4)
+    res = Dict{Int64, Agent1}(4 => Agent1(4, (2, 2)), 2 => Agent1(2, (2, 2)),
+                              3 => Agent1(3, (2, 2)), 1 => Agent1(1, (1, 1)))
+    res_fields = [getfield(res[k], f) for f in fieldnames(Agent1) for k in keys(res)]
+    agents_fields = [getfield(a, f) for f in fieldnames(Agent1) for a in allagents(model4)]
+    @test allids(model4) == keys(res)
+    @test res_fields == agents_fields
+    sample!(model4, 2)
+    res = Dict{Int64, Agent1}(4 => Agent1(4, (2, 2)), 1 => Agent1(1, (1, 1)))
+    res_fields = [getfield(res[k], f) for f in fieldnames(Agent1) for k in keys(res)]
+    agents_fields = [getfield(a, f) for f in fieldnames(Agent1) for a in allagents(model4)]
+    @test allids(model4) == keys(res)
+    @test res_fields == agents_fields
 
-        rng = StableRNG(42)
-        model = StandardABM(Agent2; rng = rng, container = c, warn_deprecation = false)
-        for i in 1:20
-            add_agent!(model, rand(abmrng(model)))
-        end
-        allweights = [i.weight for i in allagents(model)]
-        mean_weights = sum(allweights) / length(allweights)
-        sample!(model, 12, :weight)
-        @test nagents(model) == 12
-        allweights = [i.weight for i in allagents(model)]
-        mean_weights_new = sum(allweights) / length(allweights)
-        @test mean_weights_new > mean_weights
-
-        model2 = StandardABM(Agent2; rng = rng, container = c, warn_deprecation = false)
-        while true
-            for i in 1:20
-                add_agent!(model2, rand(abmrng(model2)) / rand(abmrng(model2)))
-            end
-            allweights = [i.weight for i in allagents(model2)]
-            allunique(allweights) && break
-        end
-        # Cannot draw 50 samples out of a pool of 20 without replacement
-        @test_throws ErrorException sample!(model2, 50, :weight; replace = false)
-        sample!(model2, 15, :weight; replace = false)
-        allweights = [i.weight for i in allagents(model2)]
-        @test allunique(allweights)
-
-        model3 = StandardABM(Agent2; rng = rng, container = c, warn_deprecation = false)
-        # Guarantee all starting weights are unique
-        while true
-            for i in 1:20
-                add_agent!(model3, rand(abmrng(model3)) / rand(abmrng(model3)))
-            end
-            allweights = [i.weight for i in allagents(model3)]
-            allunique(allweights) && break
-        end
-        sample!(model3, 100, :weight; replace = true)
-        allweights = [i.weight for i in allagents(model3)]
-        @test !allunique(allweights)
+    rng = StableRNG(42)
+    model = StandardABM(Agent2; rng = rng, warn_deprecation = false)
+    for i in 1:20
+        add_agent!(model, rand(abmrng(model)))
     end
+    allweights = [i.weight for i in allagents(model)]
+    mean_weights = sum(allweights) / length(allweights)
+    sample!(model, 12, :weight)
+    @test nagents(model) == 12
+    allweights = [i.weight for i in allagents(model)]
+    mean_weights_new = sum(allweights) / length(allweights)
+    @test mean_weights_new > mean_weights
+
+    model2 = StandardABM(Agent2; rng = rng, warn_deprecation = false)
+    while true
+        for i in 1:20
+            add_agent!(model2, rand(abmrng(model2)) / rand(abmrng(model2)))
+        end
+        allweights = [i.weight for i in allagents(model2)]
+        allunique(allweights) && break
+    end
+    # Cannot draw 50 samples out of a pool of 20 without replacement
+    @test_throws ErrorException sample!(model2, 50, :weight; replace = false)
+    sample!(model2, 15, :weight; replace = false)
+    allweights = [i.weight for i in allagents(model2)]
+    @test allunique(allweights)
+
+    model3 = StandardABM(Agent2; rng = rng, warn_deprecation = false)
+    # Guarantee all starting weights are unique
+    while true
+        for i in 1:20
+            add_agent!(model3, rand(abmrng(model3)) / rand(abmrng(model3)))
+        end
+        allweights = [i.weight for i in allagents(model3)]
+        allunique(allweights) && break
+    end
+    sample!(model3, 100, :weight; replace = true)
+    allweights = [i.weight for i in allagents(model3)]
+    @test !allunique(allweights)
 end
 
 @testset "random agent" begin