This is possible, and even much faster than when using a Dict (wrong)