@Tortar 2023-12-12
Actually I don't really like that the same id will be given to different agents when using `sample!` with a vector