@Tortar 2023-12-14
As you can see, the speed-up is very good: 

```
# this pr:

@btime sample!(model, 10^6) setup=(model = fake_model(Xoshiro(42), 10^7)) evals=1 samples=10 seconds=10
276.573 ms (1000010 allocations: 114.44 MiB)

# before:

@btime sample!(model, 10^6) setup=(model = fake_model(Xoshiro(42), 10^7)) evals=1 samples=10 seconds=10
  1.947 s (48424 allocations: 168.59 MiB)
```

(I changed the scope of this pr to be just an optimization of the current function)