@Tortar 2023-12-15
Actually I used a model with no space to benchmark, the new method falls behind when there is a space instead. Need to change the code to acknowledge this