@Tortar 2023-12-21
should be okay now, this is an example benchmark:

```julia
using Agents, BenchmarkTools, Random

# no space model

model_step!(model) = nothing

function fake_model(rng, nagents)
    model = StandardABM(NoSpaceAgent; model_step!, rng)
    for i in 1:nagents
        add_agent!(model)
    end
    model
end

@btime sample!(model, 10^5) setup=(model = fake_model(Xoshiro(42), 10^6)) evals=1

# space model

function fake_model(rng, nagents)
    model = StandardABM(GridAgent{2}, GridSpace((100, 100)); model_step!, rng)
    for i in 1:nagents
        add_agent!(model)
    end
    model
end

@btime sample!(model, 10^5) setup=(model = fake_model(Xoshiro(42), 10^6)) evals=1
```

where we have that

```julia
before pr, no space: 87.246 ms (4769 allocations: 16.87 MiB)
after pr, no space: 17.157 ms (100010 allocations: 11.45 MiB)

before pr, no space: 162.883 ms (4789 allocations: 16.95 MiB) 
after pr, no space: 22.714 ms (100010 allocations: 12.98 MiB)
```

In general, the perf improvement can vary, but it should be always an improvement except for edge cases like 10 agent in a 1 million positions space.